requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
forex-python==1.6
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-decouple==3.4
requests==2.25.1
simplejson==3.17.2
urllib3==1.26.6
Werkzeug==2.0.1
